fix: :bug: Use i64 for nonce

+++ client/blockchain-service-db/src/models.rs
@@ -8,7 +8,7 @@ use crate::schema::pending_transactions;
 #[diesel(primary_key(account_id, nonce))]
 pub struct PendingTransactionRow {
     pub account_id: Vec<u8>,
-    pub nonce: i32,
+    pub nonce: i64,
     pub hash: Vec<u8>,
     pub call_scale: Vec<u8>,
     pub state: String,
@@ -21,11 +21,9 @@ pub struct PendingTransactionRow {
 #[diesel(table_name = pending_transactions)]
 pub struct NewPendingTransaction<'a> {
     pub account_id: &'a [u8],
-    pub nonce: i32,
+    pub nonce: i64,
     pub hash: &'a [u8],
     pub call_scale: &'a [u8],
     pub state: &'a str,
     pub creator_id: &'a str,
 }
-
-
